refactor decision/executor io
makes deciders return decisions rather than executors, so the fallback decision is more transparant to the twitch decicder

Still need to fix decider_twitch.go, this doesn't compile.

Make it so the fallback runs to start with, and that is used as the computer vote in the vote status broadcast. And the tallying systems get initialised to that default vote.

we'll need to keep track of oid for voters to prevent duplicate votes in future, but just focus on fixing the errors for now